fixed bugs

+++ Day-2.py
@@ -1,8 +1,8 @@
 # Q1. Bill calculator
 
-# Product_1 = int(input("Enter the price of product 1: "))
-# Product_2 = int(input("Enter the price of product 2: "))
-# Product_3 = int(input("Enter the price of product 3: "))
+# Product_1 = float(input("Enter the price of product 1: "))
+# Product_2 = float(input("Enter the price of product 2: "))
+# Product_3 = float(input("Enter the price of product 3: "))
 
 # Sum = Product_1 + Product_2 + Product_3
 # Discounted_bill = Sum - 0.20 * Sum
@@ -13,7 +13,7 @@
 
 # Q2. Calculate gross salary
 
-Basic_salary = int(input("Enter the basic salary: "))
+Basic_salary = float(input("Enter the basic salary: "))
 HRA = 0.20 * Basic_salary
 DA = 0.15 * Basic_salary
 Gross_salary = Basic_salary + HRA + DA
